% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_europe_map.R
\docType{data}
\name{europe_map_nuts3}
\alias{europe_map_nuts3}
\title{European Map (NUTS3)}
\format{
\describe{
\item{long}{Longitude.}
\item{lat}{Latitude.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{nuts3}{2 character NUTS3 code, describing administrative regions defined by Eurostat/GISCO.}
\item{location name}{Location name}
\item{country_code}{2 character NUTS0 code, describing the country}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts}
}
\usage{
europe_map_nuts3
}
\description{
We conveniently package map datasets for Europe
(taken from Eurostat/GISCO) that can be used in ggplot2 without needing any geo
libraries. This data is licensed under specific conditions noted in the details section.
}
\details{
The permission to use the data is granted on condition that:
\itemize{
\item The data will not be used for commercial purposes
\item The source will be acknowledged. A copyright notice will have to be visible on any printed or electronic publication using this dataset
}
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat, group = group))
q <- q + geom_polygon(data = fhidata::europe_map_nuts3, color = "black", fill = "white", size = 0.2)
q <- q + theme_void()
q <- q + coord_map(
  projection = "cylequalarea",
  xlim = c(-25, 50),
  ylim = c(31, 75),
  orientation = c(90, 0, 0),
  parameters = list(lat0 = 90)
)
q <- q + labs(caption = "\u00A9 EuroGeographics for the administrative boundaries")
q
}
\keyword{datasets}
