% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_lau2_map_b2019_default_dt}
\alias{norway_lau2_map_b2019_default_dt}
\title{Map of Norwegian municipalities (2019 borders)}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{location_code}{Location code (municipality code).}
}
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/uuid/cb02ab77-d3e6-4500-8a92-ea67367e7734}
}
\usage{
norway_lau2_map_b2019_default_dt
}
\description{
We conveniently package map datasets for Norwegian municipalities
(taken from Geonorge) that can be used in ggplot2 without needing any geo
libraries. This data is licensed under Creative Commons 0.
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhimaps::norway_lau2_map_b2019_default_dt,
  mapping = aes(group = group),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
