% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_lau2_map_b2019_default_sf}
\alias{norway_lau2_map_b2019_default_sf}
\title{Map of Norwegian municipalities (2019 borders) in sf format}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{location_code}{Location code (municipality code).}
}
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/uuid/cb02ab77-d3e6-4500-8a92-ea67367e7734}
}
\usage{
norway_lau2_map_b2019_default_sf
}
\description{
This data is licensed under Creative Commons 0.
}
\keyword{datasets}
