% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_lau2_map_b2020_default_dt}
\alias{norway_lau2_map_b2020_default_dt}
\alias{norway_lau2_position_geolabels_b2020_default_dt}
\title{Map of Norwegian municipalities (2020 borders) in data.table format}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{location_code}{Location code (municipality code).}
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 356 rows and 3 columns.
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/norske-fylker-og-kommuner-illustrasjonsdata-2020-(klippet-etter-kyst)/7408853f-eb7d-48dd-bb6c-80c7e80f7392}
}
\usage{
norway_lau2_map_b2020_default_dt

norway_lau2_position_geolabels_b2020_default_dt
}
\description{
We conveniently package map datasets for Norwegian municipalities
(taken from Geonorge) that can be used in ggplot2 without needing any geo
libraries. This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhimaps::norway_lau2_map_b2020_default_dt,
  mapping = aes(group = group),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
