% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_map_county_label_positions_b2020}
\alias{norway_map_county_label_positions_b2020}
\title{Label Positions for Norwegian county (2020 borders)}
\format{
\describe{
\item{location_code}{Location code (county code).}
\item{long}{Location code.}
\item{lat}{Location name.}
}
}
\usage{
norway_map_county_label_positions_b2020
}
\description{
GPS coordinates for labels for Norwegian county
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhimaps::norway_map_county_b2020,
  mapping = aes(group = group),
  color = "black",
  fill = "white"
)
q <- q + geom_text(
  data = fhimaps::norway_map_county_label_positions_b2020,
  mapping = aes(label = location_code),
  color = "red"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
