% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_nuts3_positions_geolabels_b2020_default_dt}
\alias{norway_nuts3_positions_geolabels_b2020_default_dt}
\title{Label Positions for Norwegian county (2020 borders)}
\format{
\describe{
\item{location_code}{Location code (county code).}
\item{long}{Location code.}
\item{lat}{Location name.}
}
}
\usage{
norway_nuts3_positions_geolabels_b2020_default_dt
}
\description{
GPS coordinates for labels for Norwegian county
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhimaps::norway_nuts3_map_b2020_default_dt,
  mapping = aes(group = group),
  color = "black",
  fill = "white"
)
q <- q + geom_text(
  data = fhimaps::norway_nuts3_positions_geolabels_b2020_default_dt,
  mapping = aes(label = location_code),
  color = "red"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
