% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_xxx_position_title_insert_oslo_b2019_insert_oslo_dt}
\alias{norway_xxx_position_title_insert_oslo_b2019_insert_oslo_dt}
\title{Title Positions for Inserts (2019 borders)}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
}
}
\usage{
norway_xxx_position_title_insert_oslo_b2019_insert_oslo_dt
}
\description{
GPS coordinates for inserts for Norwegian maps
}
\examples{
library(ggplot2)
pd <- fhimaps::norway_nuts3_map_b2019_insert_oslo_dt
q <- ggplot()
q <- q + geom_polygon(
  data = pd,
  aes(x = long, y = lat, group = group, fill = location_code),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + annotate(
  "text",
  x = fhimaps::norway_xxx_position_title_insert_oslo_b2019_insert_oslo_dt$long,
  y = fhimaps::norway_xxx_position_title_insert_oslo_b2019_insert_oslo_dt$lat,
  label = "Oslo"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
