% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bydeloslo_map.R
\docType{data}
\name{oslo_ward_positions_geolabels_b2020_default_dt}
\alias{oslo_ward_positions_geolabels_b2020_default_dt}
\title{Label Positions for Oslo wards (2020 borders)}
\format{
\describe{
\item{location_code}{Location code (ward code).}
\item{long}{Location code.}
\item{lat}{Location name.}
}
}
\usage{
oslo_ward_positions_geolabels_b2020_default_dt
}
\description{
GPS coordinates for labels for Oslo wards
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhimaps::oslo_ward_map_b2020_default_dt,
  mapping = aes(group = group),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + geom_text(
  data = fhimaps::oslo_ward_positions_geolabels_b2020_default_dt,
  mapping = aes(label = location_code),
  color = "red"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
