% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_senorge.R, R/data_norway_thredds.R
\docType{data}
\name{senorge_b2019}
\alias{senorge_b2019}
\title{1x1km grids to Norwegian municipalities (2019 borders)}
\format{
\describe{
\item{row}{The row of the [X,Y,time] matrix.}
\item{col}{The column of the [X,Y,time] matrix.}
\item{location_code}{Municipality code that this 1x1km grid belongs to.}
\item{year}{The year that location_code corresponds to.}
}

\describe{
\item{row}{The row of the [X,Y,time] matrix.}
\item{col}{The column of the [X,Y,time] matrix.}
\item{location_code}{Municipality code that this 1x1km grid belongs to.}
\item{year}{The year that location_code corresponds to.}
}
}
\source{
\url{http://thredds.met.no/thredds/catalog/senorge/seNorge_2018/Archive/catalog.html}

\url{http://thredds.met.no/thredds/catalog/senorge/seNorge_2018/Archive/catalog.html}
}
\usage{
senorge_b2019

senorge_b2019
}
\description{
We classify the 1x1km grids from senorge into municipalities (2019 borders).

We classify the 1x1km grids from senorge into municipalities (2019 borders).
}
\keyword{datasets}
