% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_nuts3_map_b2020_default_sf}
\alias{norway_nuts3_map_b2020_default_sf}
\title{Map of Norwegian counties (2020 borders) in sf format}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{location_code}{Location code (county code).}
}
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/norske-fylker-og-kommuner-illustrasjonsdata-2020-(klippet-etter-kyst)/7408853f-eb7d-48dd-bb6c-80c7e80f7392}
}
\usage{
norway_nuts3_map_b2020_default_sf
}
\description{
This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\keyword{datasets}
