% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bydeloslo_map.R
\docType{data}
\name{oslo_ward_map_b2020_default_sf}
\alias{oslo_ward_map_b2020_default_sf}
\title{Maps of Oslo bydels (wards) (2020 borders) in sf format}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{location_code}{Location code (ward code).}
}
}
\usage{
oslo_ward_map_b2020_default_sf
}
\description{
This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\keyword{datasets}
